requirements.txt
#Data streaming
pika == 1.3.1
pika-stubs == 0.1.3

#ORM
sqlalchemy-migrate == 0.13.0
pymysql == 1.0.2
mysql-connector-python == 8.0.32
python-dotenv == 1.0.0

#Other packages
python-dotenv == 1.0.0
tqdm==4.65.0
requests==2.28.2